Morph indexData bytearray into proper model.
GeometryExtension used to have a custom geometryChanged signal, that
passed on a QByteArray containing the data. That signal needed ugly
decoding on client side and especially wouldn't be fired when first
opening the Geometry tab. Thus now we do the decoding on server side
and feed a proper QAIM from it.